intelligent_secure_capstone-main/ai_server/requirements.txt
tensorflow==2.19.0
keras==3.9.1
numpy==1.26.4
joblib==1.4.2
scapy==2.6.1
mysql-connector-python==9.3.0
scikit-learn==1.6.1
